Fix compatibility with vscode extension.
This should have no functional differences. This just moves some of the abstraction out of the function classes.
